# api/openapi.yaml
openapi: 3.1.0
info:
  title: TOT API
  version: 1.0.0-draft
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    JSON API for the Next.js frontend (`frontend-tot`). Draft contract — see
    `docs/frontend/inventory.md` for the page inventory and open decisions.

    Conventions:
    - Base path `/api/v1`; frontend and API share one origin (reverse proxy), so the
      session cookie stays `HttpOnly; SameSite=Lax` and no CORS is needed.
    - JSON bodies and fields are `snake_case`; timestamps are RFC 3339.
    - Errors always use the `Error` schema with a stable machine `code`; user-facing
      text is localized on the frontend by `code`, `message` is for logs only.
    - Unauthenticated requests get `401`, never a redirect; malformed JSON gets `400 malformed_json`.
    - Mutating requests (POST/PUT/PATCH/DELETE) must come from an allowed origin
      (`Origin` / `Sec-Fetch-Site`), otherwise `403 csrf_rejected`.
    - Request bodies are `application/json` unless stated otherwise, otherwise `415 unsupported_media_type`.
    - HTML fields (`*_html`) are rendered and sanitized server-side.
    - Admin endpoints need role `author` or `admin`; admin-only ones say so. A course
      the author cannot access answers `404`; a missing permission on an accessible
      course answers `403 forbidden` (see `CourseAccess`).
servers:
  - url: /api/v1
...
security:
  - session: []
paths:
...
  /admin/users:
    get:
      tags: [admin-users]
      summary: List users
      operationId: adminListUsers
      description: Admin only.
      parameters:
        - name: q
          in: query
          description: Substring match on name or email.
          schema: { type: string }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Users, newest first.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/AdminUser' }
                  next_cursor: { type: [string, 'null'] }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [admin-users]
      summary: Create user
      operationId: adminCreateUser
      description: Admin only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, role]
              properties:
                name: { type: string, minLength: 1, maxLength: 100 }
                email: { type: string, format: email }
                password: { type: string, minLength: 6 }
                role: { $ref: '#/components/schemas/Role' }
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminUser' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: '`email_taken`.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: session
      description: HttpOnly session cookie set by `POST /auth/login`.
  parameters:
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
...
  responses:
    Unauthorized:
      description: '`unauthorized` — no or expired session.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: '`forbidden` or `csrf_rejected`.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ValidationFailed:
      description: '`validation_failed` with per-field codes in `details.fields`.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Role:
      type: string
      enum: [student, author, admin]
      description: '`author` manages courses they own or co-author; `admin` manages everything.'
...
    Error:
      type: object
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ErrorBody' }
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          examples: [unauthorized, forbidden, not_found, validation_failed, slug_taken, already_author, sandbox_disabled]
        message:
          type: string
          description: Developer-facing, not localized.
        details: { $ref: '#/components/schemas/ErrorDetails' }
    ErrorDetails:
      type: object
      properties:
        fields:
          type: object
          additionalProperties: { type: string }
          description: Field name → error code (`required`, `too_short`, `too_long`, `invalid_format`, `invalid_value`, `not_found`, `not_author`).
        issues:
          type: array
          items: { $ref: '#/components/schemas/ImportIssue' }
...
    AdminUser:
      type: object
      required: [id, name, email, role, blocked, created_at, is_self, role_pinned_by_env]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/Role' }
        blocked: { type: boolean }
        created_at: { type: string, format: date-time }
        is_self: { type: boolean }
        role_pinned_by_env: { type: boolean, description: Email is in ADMIN_EMAILS. }

    # ── import / export ──
...
    ImportIssue:
      type: object
      required: [level, code]
      properties:
        level: { type: string, enum: [error, warn] }
        code:
          type: string
          examples: [lesson_slug_empty, lesson_slug_duplicate, no_lessons, lesson_title_empty, unknown_kind, unknown_format, too_few_options, correct_out_of_range, check_without_image]
        lesson_slug: { type: string }
        path: { type: string, description: 'JSON pointer, e.g. `/lessons/2/quiz/0`.' }
...